Splitting up the code into separate files and removing unused code.
