+++ FLEX-0S9-Net.csproj
@@ -47,8 +47,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="OS9_ID_SECTOR.cs" />
+    <Compile Include="Ports.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RAW_SIR.cs" />
+    <Compile Include="UniFLEX_SIR.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
